Refactor backend API to introduce centralized routing, authentication and error handling middleware, dedicated handlers, and a new configuration structure.
